### 2025-03-29 8fcc591
added zero changepoint option
