+++ gbnet/models/forecasting.py
@@ -22,7 +22,7 @@ def loadModule(module):
 
 
 def pd_datetime_to_seconds(x: pd.Series):
-    return pd.to_numeric(x) // 10**9
+    return pd.to_numeric(x) / 1000000000.0
 
 
 class Forecast(BaseEstimator, RegressorMixin):
@@ -87,6 +87,17 @@ class Forecast(BaseEstimator, RegressorMixin):
         changepoint_penalty=0.0,
         pytorch_lr=0.01,
     ):
+        """
+        best parameters seem to be around this while using GBLinear trend_type
+        {
+            # 'nrounds': 50,
+            # 'min_hess': 0.5,
+            # 'lambd': 0.5,
+            # 'n_changepoints': 2,
+            # 'changepoint_penalty': 1.5,
+            # 'lr': 0.7
+        }
+        """
         if params is None:
             params = {}
         self.nrounds = nrounds
@@ -126,16 +137,15 @@ class Forecast(BaseEstimator, RegressorMixin):
             optimizer.zero_grad()
             t, p, c = self.model_(df, components=True)
             loss = mse((t + p).flatten(), torch.Tensor(df["y"].values).flatten())
-            # Add L1 penalty for changepoints
+            # Add penalty for changepoints
+
             if self.trend_type == "PyTorch":
                 changepoint_penalty = self.changepoint_penalty * torch.norm(
                     self.model_.trend_changepoints.weight, p=1
                 )
             else:
-                changepoint_penalty = self.changepoint_penalty * torch.mean(
-                    torch.square(c)
-                )
-
+                penalty = self.changepoint_penalty if self.n_changepoints > 0 else 0
+                changepoint_penalty = penalty * torch.mean(torch.square(c))
             loss = loss + changepoint_penalty
 
             loss.backward(create_graph=True)
@@ -229,7 +239,7 @@ class ForecastModule(torch.nn.Module):
         self.n_changepoints = n_changepoints
 
         # Create equally spaced changepoints
-        if changepoint_range is not None:
+        if changepoint_range is not None and self.n_changepoints > 0:
             min_date, max_date = changepoint_range
             self.changepoint_locs = torch.linspace(
                 min_date, max_date, n_changepoints + 2
@@ -239,13 +249,15 @@ class ForecastModule(torch.nn.Module):
 
         if trend_type == "PyTorch":
             self.trend = torch.nn.Linear(1, 1)
-            self.trend_changepoints = torch.nn.Linear(n_changepoints, 1, bias=False)
+            self.trend_changepoints = torch.nn.Linear(
+                n_changepoints, 1, bias=False
+            )  # TODO make any fixes here that are applied to the other
             self.bn = nn.LazyBatchNorm1d()
             self.bn_changepoints = nn.LazyBatchNorm1d()
         if trend_type == "GBLinear":
             self.trend = GBLinear(1, 1, **gblinear_params)
             self.trend_changepoints = GBLinear(
-                n_changepoints, 1, bias=False, **gblinear_params
+                max(n_changepoints, 1), 1, bias=False, **gblinear_params
             )
 
         GBModule = loadModule(module_type)
@@ -274,14 +286,15 @@ class ForecastModule(torch.nn.Module):
         torch.Tensor
             Normalized changepoint features of shape [n_samples, n_changepoints]
         """
+        ncp = max(self.n_changepoints, 1)
+
         if self.changepoint_locs is None:
-            return torch.zeros((numeric_dt.shape[0], self.n_changepoints))
+            return torch.zeros((numeric_dt.shape[0], ncp))
 
         # ReLU function for changepoints: max(0, t - changepoint_loc)
         features = torch.maximum(
-            torch.zeros_like(numeric_dt.expand(-1, self.n_changepoints)),
[…]
