### 2025-03-30 d31d4f2
add gbtrend
