+++ gbnet/models/forecasting.py
@@ -9,6 +9,10 @@ from sklearn.utils.validation import check_is_fitted
 from gbnet.gblinear import GBLinear
 
 
+def pd_datetime_to_seconds(x: pd.Series):
+    return pd.to_numeric(x) / 1000000000.0
+
+
 def loadModule(module):
     assert module in {"XGBModule", "LGBModule"}
     if module == "XGBModule":
@@ -187,6 +191,11 @@ class ForecastModule(torch.nn.Module):
             output_dim=1,
             params=params,
         )
+        trend_fn_params = params.copy()
+        trend_fn_params["eta"] = trend_fn_params["eta"] / 2
+        self.trend_fn = GBModule(
+            batch_size=n, input_dim=1, output_dim=1, params=trend_fn_params
+        )
         self.initialized = False
         self.trend_type = trend_type
 
@@ -240,7 +249,7 @@ class ForecastModule(torch.nn.Module):
         df["hour"] = df["ds"].dt.hour
         df["minute"] = df["ds"].dt.minute
         df["weekday"] = df["ds"].dt.weekday
-        df["numeric_dt"] = pd.to_numeric(df["ds"])
+        df["numeric_dt"] = pd_datetime_to_seconds(df["ds"])
         if not self.initialized:
             self.initialize(df)
 
@@ -249,15 +258,23 @@ class ForecastModule(torch.nn.Module):
         if self.trend_type == "PyTorch":
             trend_component = self.trend(
                 self.bn(
-                    torch.Tensor(np.array(pd.to_numeric(df["ds"]))).reshape([-1, 1])
+                    torch.Tensor(np.array(pd_datetime_to_seconds(df["ds"]))).reshape(
+                        [-1, 1]
+                    )
                 )
             )
         if self.trend_type == "GBLinear":
             trend_component = self.trend(
-                torch.Tensor(np.array(pd.to_numeric(df["ds"]))).reshape([-1, 1])
+                torch.Tensor(np.array(pd_datetime_to_seconds(df["ds"]))).reshape(
+                    [-1, 1]
+                )
             )
 
-        forecast = trend_component + self.periodic_fn(X)
+        forecast = (
+            trend_component
+            + self.periodic_fn(X)
+            + self.trend_fn(df["numeric_dt"]) * df["numeric_dt"]
+        )
 
         return forecast
 
@@ -266,188 +283,3 @@ class ForecastModule(torch.nn.Module):
             self.trend.gb_step()
 
         self.periodic_fn.gb_step()
-
-
-class NSForecast(BaseEstimator, RegressorMixin):
-    """
-    A forecasting model class that implements a trend + seasonality + minor non-stationarity
-    model using XGBModule or LGBModule
-
-    Parameters
-    ----------
-    nrounds : int, default=3000
-        Number of training iterations (epochs) for the model. Overfitting seems to
-        be fine usually.
-    xcols : list, optional
-        List of column names to be used as input features for the model.
-    params : dict, optional
-        Dictionary of additional parameters to be passed to the underlying forecast model.
-
-    Attributes
-    ----------
-    nrounds : int
-        Number of training rounds for the model.
-    xcols : list
-        Input feature column names.
-    params : dict
-        Additional parameters passed to the forecast model.
-    model_ : NSForecastModule or None
-        Trained forecast model instance. Set after fitting.
-    losses_ : list
-        List of loss values recorded at each training iteration.
-
-    Methods
-    -------
-    fit(X, y)
-        Trains the forecast model using the input features X and target variable y.
-        X must contain the datetime column 'ds'.
-    predict(X)
-        Predicts target values based on the input features X.
-
-    Notes
-    -----
[…]
