### 2025-04-02 #7934 @osorensen
:wave: @mrava87 @harishsiravuri @TerenceLiu98 @hanyyi @jameshorine would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
