### 2025-04-08 c22a488
added train gap param
