### 2025-04-10 1c13f6e
removed trend type
