+++ gbnet/models/tests/test_forecasting.py
@@ -4,6 +4,7 @@ from datetime import datetime, timedelta
 import numpy as np
 import pandas as pd
 from gbnet.models import forecasting
+import torch
 
 
 class TestForecast(unittest.TestCase):
@@ -127,3 +128,208 @@ class TestForecast(unittest.TestCase):
             all(x >= y for x, y in zip(losses[:10], losses[1:11])),
             "Losses are not monotonically decreasing.",
         )
+
+    def test_changepoint_initialization(self):
+        """
+        Test that changepoints are properly initialized with the correct number and spacing.
+        """
+        # Initialize estimator with changepoint parameters
+        estimator = forecasting.Forecast(
+            nrounds=100,
+            changepoint_params={
+                "n_changepoints": 5,
+                "cp_gap": 0.8,
+                "cp_train_gap": 2,
+            },
+        )
+
+        # Fit the model
+        estimator.fit(self.df[["ds"]], self.df["y"])
+
+        # Check that changepoints were initialized
+        self.assertTrue(
+            hasattr(estimator.model_, "cp_input"), "Changepoints not initialized"
+        )
+        self.assertEqual(
+            len(estimator.model_.cp_input), 5, "Incorrect number of changepoints"
+        )
+
+        # Check changepoint spacing
+        cp_times = estimator.model_.cp_input.flatten()
+        expected_min = self.df["ds"].min().timestamp()
+        expected_max = self.df["ds"].min().timestamp() + 0.8 * (
+            self.df["ds"].max().timestamp() - self.df["ds"].min().timestamp()
+        )
+        self.assertGreaterEqual(
+            cp_times[0], expected_min, "First changepoint too early"
+        )
+        self.assertLessEqual(cp_times[-1], expected_max, "Last changepoint too late")
+
+    def test_changepoint_prediction(self):
+        """
+        Test that predictions with changepoints work correctly and return expected components.
+        """
+        # Initialize estimator with changepoints
+        estimator = forecasting.Forecast(
+            nrounds=100, changepoint_params={"n_changepoints": 3}
+        )
+
+        # Fit the model
+        estimator.fit(self.df[["ds"]], self.df["y"])
+
+        # Get predictions with components
+        trend, periodic, changepoints = estimator.predict(
+            self.df[["ds"]], components=True
+        )
+
+        # Check component shapes
+        self.assertEqual(len(trend), len(self.df), "Trend component length mismatch")
+        self.assertEqual(
+            len(periodic), len(self.df), "Periodic component length mismatch"
+        )
+        self.assertEqual(
+            len(changepoints), len(self.df), "Changepoint component length mismatch"
+        )
+
+        # Check that components are not NaN
+        self.assertFalse(
+            np.isnan(trend.detach().numpy()).any(),
+            "Trend component contains NaN values",
+        )
+        self.assertFalse(
+            np.isnan(periodic.detach().numpy()).any(),
+            "Periodic component contains NaN values",
+        )
+        self.assertFalse(
+            np.isnan(changepoints.detach().numpy()).any(),
+            "Changepoint component contains NaN values",
+        )
+
+    def test_changepoint_impact(self):
+        """
+        Test that changepoints have a meaningful impact on predictions.
+        """
+        # Create a dataset with a clear changepoint
+        dates = pd.date_range(start="2020-01-01", periods=200, freq="D")
+        y = np.concatenate(
+            [
+                np.linspace(0, 1, 100),  # First trend
[…]
