### 2025-04-13 0f41aa6
try just outputing a variance
