### 2025-04-13 #7934 @animikhaich
Hi @osorensen Thanks for the invitation. I'm willing to review this submission for JOSS. Thanks!
