### 2025-04-13 #8047 @editorialbot
[REVIEW]: GBNet: Gradient Boosting packages integrated into PyTorch
**Submitting author:** @mthorrell (<a href="http://orcid.org/0009-0001-3091-0342">Michael Horrell</a>)
**Repository:** https://github.com/mthorrell/gbnet
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.3.0
