### 2025-04-23 f55191c
added linear reg uncertainty
