### 2025-05-05 #8047 @animikhaich
Hi @osorensen I am unable to tag the author (@mthorrell). I have a few comments regarding the documentation and implementation. Anything I can do to tag the author?

Comments:
1. The official documentation seems incomplete outside in-code doc strings. That needs a revision and update: https://gbnet.readthedocs.io/en/latest/index.html
2. The `requirements.txt` file is incomplete. Upon trying to run the example usage from the [Forecasting section of the Readme](https://github.com/mthorrell/gbnet?tab=readme-ov-file#forecasting) I got the following errors.:
```py
Traceback (most recent call last):
  File "/home/ani/Projects/gbnet/test.py", line 198, in <module>
    from prophet import Prophet
ModuleNotFoundError: No module named 'prophet'

Importing plotly failed. Interactive plots will not work.
````
The following dependencies need to be added to the `requirements_test.txt`:
```sh
plotly
prophet
```

3. Upon trying to run the [first example code snippet from the README](https://github.com/mthorrell/gbnet?tab=readme-ov-file#basic-training-of-a-gbm-for-comparison-to-existing-gradient-boosting-packages), I got the following warning - this is not a critical warning. However, ideally, warnings should be fixed prior to release:
```sh
/home/ani/miniconda3/envs/review/lib/python3.12/site-packages/torch/autograd/graph.py:824: UserWarning: Using backward() with create_graph=True will create a reference cycle between the parameter and its gradient which can cause a memory leak. We recommend using autograd.grad when creating the graph to avoid this. If you have to use this function, make sure to reset the .grad fields of your parameters to None after use to break the cycle and avoid the leak. (Triggered internally at /pytorch/torch/csrc/autograd/engine.cpp:1273.)
  return Variable._execution_engine.run_backward(  # Calls into the C++ engine to run the backward pass
```
