### 2025-05-06 444238e
added example dependencies
