### 2025-05-06 #8047 @mthorrell
@animikhaich , my initial responses:
1. I will work on the official documentation, moving more of the readme.md items into the website.
2. Rather than add to `requirements_test.txt` which stores requirements for the tests, I added a new `requirements_examples.txt` with the `prophet` dependency. I also added an `examples/README.md` that indicates to users that, to run all examples, they should install requirements in the `requirements_examples.txt`.  I'm happy to also add `plotly` if you think needed, but it isn't used in the example code at all, so it seems an unnecessary import.
3. Unfortunately this warning is expected (although it is benign due to zero-ing out gradients), and I think it is mostly unavoidable using the latest PyTorch version. PyTorch creates this warning only on the first time that `backward(create_graph=True)` is called.  `create_graph=True` is required to calculate hessian information.  And using the loss.backward(...) command is standard PyTorch usage.  It might be possible to try to provide an alternative version of `backward` that uses `autograd.grad` under the hood, but I think this may provide a worse user experience. Please let me know any feedback on this.

For (2), you should be able to see updates in the Paper PR (https://github.com/mthorrell/gbnet/pull/65)
