### 2025-05-08 #8047 @animikhaich
@mthorrell, thanks for your response.

1. Please update us here once the documentation is updated.
2. The PR is satisfactory.
3. I see, thanks for explaining the root cause of the warning. You are correct about using `autograd.grad` will likely provide a worse user experience, and I tried a few things, but there does not seem to be a way around that warning. So, I believe we are good on that.

I'll proceed with the review. Thanks!
