+++ gbnet/models/tests/test_forecasting.py
@@ -24,22 +24,20 @@ class TestForecast(unittest.TestCase):
         test_df = self.df.iloc[150:]
 
         # Initialize and fit the estimator
-        estimator = forecasting.Forecast(nrounds=100)
+        estimator = forecasting.Forecast(nrounds=10)
         estimator.fit(train_df[["ds"]], train_df["y"])
 
         # Make predictions
         predictions = estimator.predict(test_df[["ds"]])
 
         # Check that predictions have the correct length
-        self.assertEqual(len(predictions), len(test_df), "Prediction length mismatch.")
+        self.assertEqual(
+            predictions.shape[0], len(test_df), "Prediction length mismatch."
+        )
 
         # Check that predictions are not NaN
-        self.assertFalse(np.isnan(predictions).any(), "Predictions contain NaN values.")
-
-        # Optionally, check that predictions are numerical
-        self.assertTrue(
-            np.issubdtype(predictions.dtype, np.number),
-            "Predictions are not numerical.",
+        self.assertFalse(
+            np.isnan(predictions).any().any(), "Predictions contain NaN values."
         )
 
     def test_fit_predict_lgbm(self):
@@ -51,22 +49,20 @@ class TestForecast(unittest.TestCase):
         test_df = self.df.iloc[150:]
 
         # Initialize and fit the estimator
-        estimator = forecasting.Forecast(nrounds=100, module_type="LGBModule")
+        estimator = forecasting.Forecast(nrounds=10, module_type="LGBModule")
         estimator.fit(train_df[["ds"]], train_df["y"])
 
         # Make predictions
         predictions = estimator.predict(test_df[["ds"]])
 
         # Check that predictions have the correct length
-        self.assertEqual(len(predictions), len(test_df), "Prediction length mismatch.")
+        self.assertEqual(
+            predictions.shape[0], len(test_df), "Prediction length mismatch."
+        )
 
         # Check that predictions are not NaN
-        self.assertFalse(np.isnan(predictions).any(), "Predictions contain NaN values.")
-
-        # Optionally, check that predictions are numerical
-        self.assertTrue(
-            np.issubdtype(predictions.dtype, np.number),
-            "Predictions are not numerical.",
+        self.assertFalse(
+            np.isnan(predictions).any().any(), "Predictions contain NaN values."
         )
 
     def test_loss_decrease(self):
@@ -77,7 +73,7 @@ class TestForecast(unittest.TestCase):
         train_df = self.df.iloc[:150]
 
         # Initialize and fit the estimator
-        estimator = forecasting.Forecast(nrounds=100)
+        estimator = forecasting.Forecast(nrounds=10)
         estimator.fit(train_df[["ds"]], train_df["y"])
 
         # Get the recorded losses
@@ -107,7 +103,7 @@ class TestForecast(unittest.TestCase):
         train_df = self.df.iloc[:150]
 
         # Initialize and fit the estimator
-        estimator = forecasting.Forecast(nrounds=100, module_type="LGBModule")
+        estimator = forecasting.Forecast(nrounds=10, module_type="LGBModule")
         estimator.fit(train_df[["ds"]], train_df["y"])
 
         # Get the recorded losses
@@ -135,7 +131,7 @@ class TestForecast(unittest.TestCase):
         """
         # Initialize estimator with changepoint parameters
         estimator = forecasting.Forecast(
-            nrounds=100,
+            nrounds=10,
             changepoint_params={
                 "n_changepoints": 5,
                 "cp_gap": 0.8,
@@ -171,39 +167,18 @@ class TestForecast(unittest.TestCase):
         """
         # Initialize estimator with changepoints
[…]
