+++ gbnet/models/tests/test_forecasting.py
@@ -1,8 +1,11 @@
 import unittest
+from unittest.mock import MagicMock, patch
+
 from datetime import datetime, timedelta
 
 import numpy as np
 import pandas as pd
+from pandas.testing import assert_series_equal, assert_frame_equal
 from gbnet.models import forecasting
 import torch
 
@@ -310,3 +313,236 @@ class TestForecast(unittest.TestCase):
         # t >= 2.0: -1.0 * (t - 1.0) - 2.0 * (t - 2.0)
         expected = torch.tensor([0.0, 0.0, -0.5, -1.0, -2.5, -4.0], dtype=torch.float32)
         self.assertTrue(torch.allclose(result, expected), "Negative slopes test failed")
+
+
+class TestMleG(unittest.TestCase):
+    def test_mle_G_when_true_G_is_zero_perfect_data(self):
+        """Test _mle_G when true G is zero (r_i^2 = S2)."""
+        S2 = 1.0
+        h = np.array([0.0, 1.0, 2.0, 3.0, 4.0, 5.0])  # Needs some non-zero h
+        r_sq = np.full_like(h, S2)
+        r = np.sqrt(r_sq)
+
+        G_estimated, _ = forecasting._mle_G(
+            r, h, S2, bracket_pad=0.01
+        )  # Small pad for precision
+        self.assertAlmostEqual(G_estimated, 0.0, places=5)
+
+    def test_mle_G_when_true_G_is_zero_h_all_zeros(self):
+        """Test _mle_G when h is all zeros, G should be 0."""
+        S2 = 1.0
+        h = np.zeros(10)
+        r = np.full(10, np.sqrt(S2))  # r^2 = S2 exactly
+
+        # If all h are zero, score function is identically zero.
+        # root_scalar might return 'a' (0.0) from bracket (a,b)
+        G_estimated, _ = forecasting._mle_G(r, h, S2, bracket_pad=0.1)
+        self.assertAlmostEqual(G_estimated, 0.0, places=5)
+
+    def test_mle_G_positive_G_perfect_fit(self):
+        """Test _mle_G with a known positive G and no noise."""
+        S2 = 1.0
+        G_true = 0.5
+        h = np.array([0.0, 1.0, 2.0, 3.0, 4.0, 5.0])
+        # r_i^2 = S2 + G_true * h_i
+        r = np.sqrt(S2 + G_true * h)
+
+        G_estimated, _ = forecasting._mle_G(r, h, S2, bracket_pad=0.01)
+        self.assertAlmostEqual(G_estimated, G_true, places=5)
+
+    def test_mle_G_positive_G_with_noise(self):
+        """Test _mle_G with positive G and noisy r."""
+        S2 = 1.0
+        G_true = 0.5
+        N = 200  # Larger N for more stable estimate with noise
+        h = np.random.rand(N) * 5
+        h[0] = 0.0  # Ensure some h can be zero
+
+        # r_i ~ N(0, variance = S2 + G_true * h_i)
+        variances = S2 + G_true * h
+        # Ensure variances are positive
+        self.assertTrue(np.all(variances > 0.00001))  # Avoid zero variance
+
+        r = np.array([np.random.normal(0, np.sqrt(v)) for v in variances])
+
+        G_estimated, _ = forecasting._mle_G(r, h, S2, bracket_pad=1.0)
+        # With noise, G_estimated won't be exact. Check if it's reasonably close.
+        # This is a statistical estimate, so large deviation is possible for small N.
+        self.assertAlmostEqual(G_estimated, G_true, delta=0.35)  # Allow some deviation
+
+    def test_mle_G_h_must_be_non_negative_assertion(self):
[…]
