### 2025-05-21 9b1fad1
updated examples
