### 2025-06-07 e856baa
include readme info
