+++ docs/source/overview.rst
@@ -1,0 +1,251 @@
+Overview
+========
+
+GBNet is a Python library that provides XGBoost and LightGBM PyTorch Modules.
+
+There are two main components of ``gbnet``:
+
+1. ``gbnet.xgbmodule``, ``gbnet.lgbmodule`` and ``gbnet.gblinear`` provide the Pytorch Modules that allow fitting of XGBoost, LightGBM and Boosted Linear models using Pytorch's computational network and differentiation capabilities.
+
+   - For example, if :math:`F(X)` is the output of an XGBoost model, you can use Pytorch to define the loss function, :math:`L(y, F(X))`. Pytorch handles the gradients of :math:`L` so, as a user, you only specify the loss function.
+   - You can also fit two (or more) boosted models together with Pytorch-supported parametric components. For instance, a recommendation prediction might look like this: :math:`\sigma(F(user) \times G(item))` where both :math:`F` and :math:`G` are separate boosting models producing embeddings of users and items respectively. ``gbnet`` makes defining and fitting such a model almost as easy as using Pytorch itself.
+
+2. ``gbnet.models`` provides specific example estimators that accomplish things that were not previously possible using only XGBoost or LightGBM. Current models:
+
+   - ``Forecast`` is a forecasting model similar in execution to Metas' Prophet algorithm. In the settings we tested, ``gbnet.models.forecasting.Forecast`` beats the performance of Meta's Prophet algorithm.
+   - ``GBOrd`` is Ordinal Regression using GBMs (both XGBoost and LightGBM supported). The complex loss function (with fitable parameters) is specified in PyTorch and put on top of either ``XGBModule`` or ``LGBModule``.
+   - Other models with plans to be integrated are time-varying Survival analysis and more with NLP.
+
+Pytorch Modules
+--------------
+
+There are currently three Pytorch Modules in ``gbnet``: ``lgbmodule.LGBModule``, ``xgbmodule.XGBModule`` and ``gblinear.GBLinear``. These create the interface between PyTorch and the boosting algorithms. LightGBM and XGBoost are wrapped in ``LGBModule`` and ``XGBModule`` respectively. ``GBLinear`` is a linear layer that is trained with boosting (rather than gradient descent) -- for some applications it trains much faster than gradient descent.
+
+Conceptually, how can Pytorch be used to fit XGBoost or LightGBM models?
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+Gradient Boosting Machines only require gradients and, for modern packages, hessians to train. Pytorch (and other neural network packages) calculates gradients and hessians. GBMs can therefore be fit as the first layer in neural networks using Pytorch.
+
+CatBoost is also supported but in an experimental capacity since the current gbnet integration with CatBoost is not as performant as the other GBDT packages.
+
+Is training a ``gbnet`` model closer to training a neural network or to training a GBM?
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
[…]
