+++ docs/source/conf.py
@@ -26,6 +26,23 @@ extensions = [
 templates_path = ["_templates"]
 exclude_patterns = []
 
+# Add these settings to control autodoc behavior
+autodoc_default_options = {
+    "members": True,
+    "member-order": "bysource",
+    "special-members": "__init__",
+    "undoc-members": True,
+    "exclude-members": "__weakref__",
+    "show-inheritance": True,
+    "no-inherited-members": True,
+}
+
+# This will prevent duplicate parameter documentation
+autodoc_typehints = "description"
+
+# This will prevent the init method from being documented separately
+autodoc_docstring_signature = True
+autodoc_preserve_defaults = True
 
 # -- Options for HTML output -------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#options-for-html-output
@@ -35,3 +52,11 @@ html_static_path = ["_static"]
 
 autoapi_dirs = ["../../gbnet"]
 autoapi_ignore = ["*/test_*.py", "*/tests/*", "*_test.py"]
+
+# Control how parameters are displayed
+napoleon_use_param = True
+napoleon_use_rtype = True
+napoleon_use_ivar = True
+
+# Control the order of documentation
+autodoc_member_order = "bysource"
