### 2025-06-16 #8047 @mthorrell
One thing I just noticed -- the image of the Ordinal Regression isn't displaying correctly in the PDF.  I will try to fix this.  Here's how it should look in the PR in the GBNet codebase: https://github.com/mthorrell/gbnet/blob/444238e86b22b7b93f5851cecc54017d04ff4769/paper/paper.md
