### 2025-06-17 09d2754
paper updates
