+++ paper/paper.bib
@@ -63,7 +63,8 @@
   volume  = {29},
   number  = {5},
   pages   = {1189--1232},
-  year    = {2001}
+  year    = {2001},
+  doi     = {10.1214/aos/1013203451}
 }
 
 @inproceedings{chen2016xgboost,
@@ -129,7 +130,8 @@
   author={Kontschieder, Peter and Fiterau, Madalina and Criminisi, Antonio and Rota Bulo, Samuel},
   booktitle={Proceedings of the IEEE international conference on computer vision},
   pages={1467--1475},
-  year={2015}
+  year={2015},
+  doi={10.1109/ICCV.2015.172}
 }
 
 @inproceedings{deepgbm,
@@ -174,5 +176,6 @@
   volume={2},
   pages={1735--1742},
   year={2006},
-  organization={IEEE}
+  organization={IEEE},
+  doi={10.1109/cvpr.2006.100}
 }
