### 2025-06-20 18f6981
updated citations
