### 2025-06-20 #8047 @mthorrell
@bahung , I believe I have addressed your comments.

@IBCHgenomic , thanks again for your comment.  I'm not fully following your comment I think. LightGBM is parallelized (see for example the `nthread` parameter).  Since GBNet calls LightGBM and/or XGBoost under the hood, it inherits the parallelism both in XGBoost and LightGBM. `nthread` can be specified as a parameter that is subsequently passed to either XGBoost or LightGBM calls.  I can add more documentation around this?
