### 2025-06-23 #8047 @IBCHgenomic
@mthorrell thank you and i was not able to see that in the reading documentation, hence asked. Does it has some functions for the density matrix or the linearization of the matrix or adjacency matrix.
