### 2025-06-26 mthorrell/gbnet#77 @bahung
[JOSS review] Github and software function
Hi @mthorrell,

This is the new issue regarding your JOSS submission [](https://github.com/openjournals/joss-reviews/issues/8047)

I installed and ran the benchmarks successfully, but there were some small difficulties during the installation, which were due to dependencies.

Please consider adding:
1. Statement of need on GitHub
2. Dependencies list
3. [Optional] Warning in using Apple Silicon Mac: These systems might load multiple, conflicting versions of the OpenMP runtime library. This is problematic because it can lead to degraded performance or, more critically, to incorrect and unpredictable results in calculations. The safest and most robust solution involves ensuring we have a single, native Apple Silicon Homebrew installation and then reinstalling all relevant Python packages that depend on OpenMP.

So that I can clear the remaining checklists:

Bests,
