### 2025-06-26 mthorrell/gbnet#77 @mthorrell
Yes good call out and thanks for the pointer. I will address.
