### 2025-06-27 b37a503
added openmp notes
