### 2025-06-28 d31c67a
adjuted readme
