### 2025-06-28 mthorrell/gbnet#77 @mthorrell
@bahung I have a PR that:
1. adds statement of need to github README and online documentation
2. adds a discussion of dependencies and troubleshooting details in the installation section.

What is best practice here? Do you want to review that PR (#78)? Or would you rather I merge it and you review?
