### 2025-06-28 a80d065
fixed two changes
