### 2025-06-29 mthorrell/gbnet#77 @bahung
Hi @mthorrell, I reviewed and it looks good to me [comment](https://github.com/mthorrell/gbnet/pull/78#issuecomment-3016177860)
