### 2025-06-29 mthorrell/gbnet#77 @bahung
Close this as the author addressed my comments.
