### 2025-06-29 aba81de
scikit-learn
This is an editorial comment on your JOSS submission https://github.com/openjournals/joss-reviews/issues/8047.

'scikit-learn' seems to be the correct spelling.
