### 2025-06-29 f838efc
Small edit
https://github.com/openjournals/joss-reviews/issues/8047
