### 2025-06-29 404e385
Figure should be referenced in text
Editorial comment on JOSS submission
https://github.com/openjournals/joss-reviews/issues/8047
