### 2025-06-29 #8047 @osorensen
@mthorrell I've made three PRs in the source repo, with minor edits of the paper.
