### 2025-06-30 #8047 @mthorrell
@osorensen , thank you for the comments. I have merged your edits into the paper branch.

And thank you @bahung for the detailed comments and review.
