### 2025-07-04 #8047 @crvernon
:wave: @mthorrell - I need you to address the following before I accept this submission for publication:

In your archive:

- [ ] The license you have posted in your archive does not match the one posted in your GitHub repository.  Please edit your Zenodo archive metadata to represent the correct license.  No new release necessary.

In your paper:

- [ ] LINE 125:  the "h" in "helmholtz" should be capitalized.  You can do this by wrapping curly brackets around characters you wish to maintain the formatting of in your bib file.
- [ ] LINE 130:  the "k" in "kaggle" should be capitalized.
- [ ] LINE 138:  the "b" in "bayes" should be capitalized.

Let me know when these changes have been made.  Thanks.
