+++ paper/paper.bib
@@ -30,7 +30,7 @@
 }
 
 @article{hinton1993autoencoders,
-  title={Autoencoders, minimum description length and Helmholtz free energy},
+  title={Autoencoders, minimum description length and {H}elmholtz free energy},
   author={Hinton, Geoffrey E and Zemel, Richard},
   journal={Advances in neural information processing systems},
   volume={6},
@@ -38,7 +38,7 @@
 }
 
 @misc{kingma2013auto,
-  title={Auto-encoding variational bayes},
+  title={Auto-encoding variational {B}ayes},
   author={Kingma, Diederik P and Welling, Max and others},
   year={2013},
   publisher={Banff, Canada}
@@ -87,7 +87,7 @@
 }
 
 @misc{kaggle2021survey,
-  title        = {2021 Kaggle Machine Learning \& Data Science Survey},
+  title        = {2021 {K}aggle Machine Learning \& Data Science Survey},
   author       = {Kaggle},
   howpublished = {Kaggle, \url{https://www.kaggle.com/c/kaggle-survey-2021}},
   year         = {2021},
